+++ lib/eda/activity.ex
@@ -2,6 +2,17 @@ defmodule EDA.Activity do
   @moduledoc "Represents a Discord presence activity."
   use EDA.Event.Access
 
+  @status_display_types %{0 => :name, 1 => :state, 2 => :details}
+
+  @types %{
+    0 => :playing,
+    1 => :streaming,
+    2 => :listening,
+    3 => :watching,
+    4 => :custom,
+    5 => :competing
+  }
+
   defstruct [
     :name,
     :type,
@@ -25,7 +36,15 @@ defmodule EDA.Activity do
 
   @type t :: %__MODULE__{
           name: String.t() | nil,
-          type: integer() | nil,
+          type:
+            :playing
+            | :streaming
+            | :listening
+            | :watching
+            | :custom
+            | :competing
+            | integer()
+            | nil,
           url: String.t() | nil,
           created_at: DateTime.t() | nil,
           timestamps: map() | nil,
@@ -39,7 +58,7 @@ defmodule EDA.Activity do
           instance: boolean() | nil,
           flags: integer() | nil,
           buttons: [map()] | nil,
-          status_display_type: non_neg_integer() | nil,
+          status_display_type: :name | :state | :details | integer() | nil,
           details_url: String.t() | nil,
           state_url: String.t() | nil
         }
@@ -48,7 +67,7 @@ defmodule EDA.Activity do
   def from_raw(raw) when is_map(raw) do
     %__MODULE__{
       name: raw["name"],
-      type: raw["type"],
+      type: EDA.Enum.name(@types, raw["type"]),
       url: raw["url"],
       created_at: EDA.Timestamp.from_unix_ms(raw["created_at"]),
       timestamps: raw["timestamps"],
@@ -62,7 +81,7 @@ defmodule EDA.Activity do
       instance: raw["instance"],
       flags: raw["flags"],
       buttons: raw["buttons"],
-      status_display_type: raw["status_display_type"],
+      status_display_type: EDA.Enum.name(@status_display_types, raw["status_display_type"]),
       details_url: raw["details_url"],
       state_url: raw["state_url"]
     }