+++ test/eda/poll_test.exs
@@ -31,7 +31,7 @@ defmodule EDA.PollTest do
       assert length(poll.answers) == 2
       assert poll.expiry == ~U[2025-06-01 00:00:00Z]
       assert poll.allow_multiselect == true
-      assert poll.layout_type == 1
+      assert poll.layout_type == :default
 
       assert {true, [%AnswerCount{id: 1, count: 10}, %AnswerCount{id: 2, count: 5}]} =
                poll.results
@@ -90,7 +90,7 @@ defmodule EDA.PollTest do
       assert poll.question == "Favorite color?"
       assert poll.duration == 24
       assert poll.allow_multiselect == false
-      assert poll.layout_type == 1
+      assert poll.layout_type == :default
       assert poll.answers == []
     end